Extract VMR patch handling into a separate file
